refactor: Move routes to own views, remove hardcoded paths
