+++ frontend/src/utils/values.ts
@@ -1,4 +1,9 @@
-export const ENDPOINTS = {
+export const API_ENDPOINTS = {
   SCALE: "/api/scale",
   SWITCHBOARD_DATA: "/api/switchboard"
+};
+
+export const CLIENT_ENDPOINTS = {
+  SWITCHBOARD: "/switchboard",
+  CENTRAL_PROD: "/centralproduction"
 };
