+++ frontend/src/views/centralProduction.tsx
