+++ frontend/src/views/switchboard.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { SwitchboardTable } from "../components/switchboardTable";
+
+export const Switchboard = () => {
+  return (
+    <div>
+      <SwitchboardTable />
+    </div>
+  );
+};
